[refactor] SheetsDownloaderWindowBase based on UI Toolkit

+++ Editor/Scripts/SheetsDownloader/ISheetsDownloaderWindow.cs
@@ -2,6 +2,6 @@
 {
     internal interface ISheetsDownloaderWindow
     {
-        void DownloadSheet(int sheetId);
+        void DownloadSheet(long sheetId);
     }
 }
